requirements.txt
# --- cli tools -------------
click >= 8.0.1

# --- math ------------------
numpy >= 1.20
scipy >= 1.7
numba >= 0.53.1

# --- data handling ---------
pandas >= 1.3

# --- image handling --------
pillow >= 8.2
easyocr >= 1.4
opencv-python >= 4.5

# --- plotting --------------
matplotlib >= 3.5
plotly >= 5.5.0

# --- machine learning ------
--find-links https://download.pytorch.org/whl/torch_stable.html
torch == 1.9.1
scikit-learn >= 1.0
# darts >= 0.15.0
u8darts[torch]   # = darts core + pytorch, hence no prophet
fastai >= 2.5.6

# --- parallel processing ---
joblib >= 1.1.0
psutil >= 5.9.2

# --- Jupyter ---------------
jupyterlab >= 3.4
ipywidgets >= 7.7

# --- other -----------------
tqdm >= 4.62

# --- formatting ------------
isort >= 5.10
black[jupyter] >= 22.3
nbstripout >= 0.6.0

# --- testing ---------------
pytest >= 6.2.0